Added audio for enemy/updated AI states
Added running/walking audio for the enemy which has audio drop off from where the player is.

Created a script that holds the audio source then controls it from that script which is accessible from the AiAgent.

Also made some fixes to the AI states and triggers for when the audio should play.